+++ packages/cp-listing-service/src/utils/data/test.ts
@@ -1,0 +1,4 @@
+import { mdiIconsList } from '@cribplug/common';
+import mdiIcons from './mdiIcons.json';
+
+console.log({ mdiIcons: mdiIcons.length, oldIcons: mdiIconsList.length });